chatagent-container/pyproject.toml

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.poetry]
name = "chatbot"
version = "0.3.0"
description = "Chat Bot for Kubernetes"
authors = ["Ben Greene <BenJGreene@gmail.com>"]
readme = "README.md"
license = "GPL3"
repository = "https://github.com/PolecatWorks/chatbot"
include = ["LICENSE"]

[tool.poetry.dependencies]
python = "^3.12"
click = "^8"
# "ruamel.yaml" = "^0.18.6"
pydantic-yaml = {extras = ["pyyaml"], version = "^1.4"}
aiohttp = "^3"
pydantic-settings = {extras = ["yaml"], version = "^2.6"}
pydantic-file-secrets = "^0.4"
prometheus-async = "^25.1.0"
langchain-openai = "^1.0"
langchain-google-genai = "^3.0"
langchain = "^1.0"
langchain-mcp-adapters = "^0.1"
langgraph = "^1.0"
grandalf = "^0.8"
mcp = "^1.21"
microsoft-agents-hosting-aiohttp = "^0.5.3"
microsoft-agents-hosting-core = "^0.5.3"
microsoft-agents-authentication-msal = "^0.5.3"
microsoft-agents-activity = "^0.5.3"


[tool.poetry.group.dev.dependencies]
aiohttp-devtools = "^1"
pytest-asyncio = "^1.1.0"
pytest = "^9"
Sphinx = "^8"
sphinx-autodoc-typehints = "^3.2.0"
pytest-sphinx = "^0.6"
flake8 = "^7.1.1"
flake8-black = "^0.4"
pylint = "^4.0"
sphinx-click = "^6.0.0"
pytest-click = "^1.1.0"
pytest-cov = "^7.0"
pytest-aiohttp = "^1"
pytest-watch = "^4.2.0"
deepeval = "^3"



[tool.poetry.scripts]
chatbot = "chatbot.cli:cli"


[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
asyncio_mode = "auto"
testpaths = ["tests"]
